day05-完成功能 （1）商品分类功能的CRUD （2）商品的简单信息录入 （3）富文本编辑器使用，获取其中内容 （4）图片上传，列表显示

+++ pinyougou_parent/pinyougou_sellergoods_service/src/main/java/com/pinyougou/sellergoods/service/impl/ItemCatServiceImpl.java
@@ -74,8 +74,12 @@ public class ItemCatServiceImpl implements ItemCatService {
 	@Override
 	public void delete(Long[] ids) {
 		for(Long id:ids){
-			itemCatMapper.deleteByPrimaryKey(id);
-		}		
+			//查询该分类下是否有数据
+			List<TbItemCat> parentItems = findByParentId(id);
+			if(parentItems==null||parentItems.size()==0){
+				itemCatMapper.deleteByPrimaryKey(id);
+			}
+		}
 	}
 	
 	
@@ -96,5 +100,15 @@ public class ItemCatServiceImpl implements ItemCatService {
 		Page<TbItemCat> page= (Page<TbItemCat>)itemCatMapper.selectByExample(example);		
 		return new PageResult(page.getTotal(), page.getResult());
 	}
-	
+
+    @Override
+    public List<TbItemCat> findByParentId(Long parentId) {
+
+		//根据上一级id查询
+		TbItemCatExample example = new TbItemCatExample();
+		Criteria criteria = example.createCriteria();
+		criteria.andParentIdEqualTo(parentId);
+		return itemCatMapper.selectByExample(example);
+	}
+
 }
